+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add "git describe" information to --version output.
+- Add keybinding help menu
 
 ## [0.5.0] - 2025-06-26