+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Add "git describe" information to --version output.
-- Add keybinding help menu
+### Added
+
+- "git describe" information in --version output.
+- Keybinding help menu.
 
 ## [0.5.0] - 2025-06-26